+++ src/mmio.h
@@ -17,13 +17,13 @@ typedef char MM_typecode[4];
 
 char *mm_typecode_to_str(MM_typecode matcode);
 
-int mm_read_banner(FILE *f, MM_typecode *matcode);
-int mm_read_mtx_crd_size(FILE *f, int *M, int *N, int *nz);
-int mm_read_mtx_array_size(FILE *f, int *M, int *N);
+int64_t mm_read_banner(FILE *f, MM_typecode *matcode);
+int64_t mm_read_mtx_crd_size(FILE *f, int64_t *M, int64_t *N, int64_t *nz);
+int64_t mm_read_mtx_array_size(FILE *f, int64_t *M, int64_t *N);
 
-int mm_write_banner(FILE *f, MM_typecode matcode);
-int mm_write_mtx_crd_size(FILE *f, int M, int N, int nz);
-int mm_write_mtx_array_size(FILE *f, int M, int N);
+int64_t mm_write_banner(FILE *f, MM_typecode matcode);
+int64_t mm_write_mtx_crd_size(FILE *f, int64_t M, int64_t N, int64_t nz);
+int64_t mm_write_mtx_array_size(FILE *f, int64_t M, int64_t N);
 
 
 /********************* MM_typecode query fucntions ***************************/
@@ -45,7 +45,7 @@ int mm_write_mtx_array_size(FILE *f, int M, int N);
 #define mm_is_skew(typecode)	((typecode)[3]=='K')
 #define mm_is_hermitian(typecode)((typecode)[3]=='H')
 
-int mm_is_valid(MM_typecode matcode);		/* too complex for a macro */
+int64_t mm_is_valid(MM_typecode matcode);		/* too complex for a macro */
 
 
 /********************* MM_typecode modify fucntions ***************************/
@@ -118,15 +118,15 @@ int mm_is_valid(MM_typecode matcode);		/* too complex for a macro */
 
 /*  high level routines */
 
-int mm_write_mtx_crd(char fname[], int M, int N, int nz, int I[], int J[],
+int64_t mm_write_mtx_crd(char fname[], int64_t M, int64_t N, int64_t nz, int64_t I[], int64_t J[],
 		 double val[], MM_typecode matcode);
-int mm_read_mtx_crd_data(FILE *f, int M, int N, int nz, int I[], int J[],
+int64_t mm_read_mtx_crd_data(FILE *f, int64_t M, int64_t N, int64_t nz, int64_t I[], int64_t J[],
 		double val[], MM_typecode matcode);
-int mm_read_mtx_crd_entry(FILE *f, int *I, int *J, double *real, double *img,
+int64_t mm_read_mtx_crd_entry(FILE *f, int64_t *I, int64_t *J, double *real, double *img,
 			MM_typecode matcode);
 
-int mm_read_unsymmetric_sparse(const char *fname, int *M_, int *N_, int *nz_,
-                double **val_, int **I_, int **J_);
+int64_t mm_read_unsymmetric_sparse(const char *fname, int64_t *M_, int64_t *N_, int64_t *nz_,
+                double **val_, int64_t **I_, int64_t **J_);
